feat(threads): implement TAS-based locking mechanism and enhance thread prioritization logging
This commit introduces a Test-And-Set (TAS) locking mechanism to improve synchronization in multi-threaded contexts. The TAS functions, including `tas_try_lock`, `tas_unlock`, and `tas_is_locked`, allow for atomic operations on lock bytes, which is essential for preventing race conditions.

Additionally, logging enhancements have been added to the `get_highest_priority_thread` function, providing better insights when threads are found or not found during scheduling decisions. These improvements aim to facilitate debugging and monitoring of thread behavior in real-time.

Key changes include:
- Implementation of TAS locking for ColdFire and standard m68k architectures.
- Replacement of integer types with short for mutex attributes to save memory.
- Additional TRACE statements to log more detailed information about thread scheduling actions.

These modifications support more efficient resource management and clearer diagnostics in the threading model.

+++ sys/proc_threads_mutex.h
@@ -25,10 +25,10 @@
 
 /* Mutex attribute structure */
 struct mutex_attr {
-    int type;                 /* Mutex type: NORMAL, RECURSIVE, ERRORCHECK */
-    int pshared;              /* Process-shared flag */
-    int protocol;             /* Priority protocol */
-    int prioceiling;          /* Priority ceiling value */
+    short type;                 /* Mutex type: NORMAL, RECURSIVE, ERRORCHECK */
+    short pshared;              /* Process-shared flag */
+    short protocol;             /* Priority protocol */
+    short prioceiling;          /* Priority ceiling value */
 };
 
 /* Mutex types */
@@ -44,12 +44,12 @@ struct mutex_attr {
 struct mutex {
     struct thread *owner;
     struct thread *wait_queue;    
-    int locked;
-    int protocol;              /* Priority protocol */
-    int type;                  /* Mutex type */
-    int prioceiling;           /* Priority ceiling */
-    int saved_priority;        /* Saved priority for ceiling protocol */
-    int lock_count;            /* Lock count for recursive mutexes */
+    short locked;
+    short protocol;              /* Priority protocol */
+    short type;                  /* Mutex type */
+    short prioceiling;           /* Priority ceiling */
+    short saved_priority;        /* Saved priority for ceiling protocol */
+    short lock_count;            /* Lock count for recursive mutexes */
 };
 
 int thread_mutex_trylock(struct mutex *mutex);
